+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ## 12.0.1 under development
 
-- no changes in this release.
+- Chg #280: Change PHP constraint in `composer.json` to `8.1 - 8.4` (@vjik)
+- Bug #280: Explicitly mark nullable parameters (@vjik)
 
 ## 12.0.0 December 23, 2024