+++ CHANGELOG.md
@@ -1,12 +1,12 @@
 # Yii View Change Log
 
-## 12.0.1 under development
+## 12.1.0 March 15, 2025
 
+- Chg #280: Change PHP constraint in `composer.json` to `8.1 - 8.4` (@vjik)
 - Enh #282: Allow using `../` in the name of the view to refer to parent directory of the directory containing the view 
   currently being rendered (@vjik)
-- Bug #282: Fix exception message when relative path is used without currently rendered view (@vjik) 
-- Chg #280: Change PHP constraint in `composer.json` to `8.1 - 8.4` (@vjik)
 - Bug #280: Explicitly mark nullable parameters (@vjik)
+- Bug #282: Fix exception message when relative path is used without currently rendered view (@vjik)
 
 ## 12.0.0 December 23, 2024